+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - ([e709cb9][]) Update react@15 to react@16
 - ([fd08633][]) Use sessión with name, rut, location, business and id course
 - ([fd08633][]) Use control to sessión
+- ([79692f9][]) Save registre with name, rut, location, business, id course, and responses.
 
 ### Fixed
 - ([8056e7d][]) Warning of react
@@ -55,3 +56,4 @@
 [ac5b4a2]: https://github.com/JonDotsoy/otecnya-questions-offline/commit/ac5b4a2
 [b771814]: https://github.com/JonDotsoy/otecnya-questions-offline/commit/b771814
 [fd08633]: https://github.com/JonDotsoy/otecnya-questions-offline/commit/fd08633
+[79692f9]: https://github.com/JonDotsoy/otecnya-questions-offline/commit/79692f9